recipe_web_app_spring  validation JS + Java
