+++ recipe_web_app_spring/src/main/java/com/example/recipe_web_app_spring/dao/IngredientDao.java
@@ -7,4 +7,6 @@ import java.util.List;
 public interface IngredientDao {
     List<Ingredient> getAllIngredientsByType(String ingredientTypeName);
 
+    List<Ingredient> getAllIngredients();
+
 }
