+++ recipe_web_app_spring/src/main/webapp/view/login.jsp
@@ -16,17 +16,28 @@
 
     <form action="login" method="POST" class="signin-form">
         <h2>Sign In</h2>
-        <input type="text" name="username" placeholder="Username" required>
-        <input type="password" name="password" placeholder="Password" required>
+        <input class="modal-input" type="text" name="username" placeholder="Username" id="username-input"
+               onblur="usernameValidation()" required>
+        <span class="error"
+              id="error-userName">Username is invalid, please enter at least 4 letters or numbers</span>
+        <input class="modal-input" type="password" name="password" placeholder="Password" id="password-input"
+               onblur="passwordValidation()" required>
+        <span class="error"
+              id="error-password">Password is invalid,please enter at least 4 letters or numbers
+                with special symbols: !, @, #, $, %</span>
         <button type="submit">Sign In</button>
+        <a href="/register">Not register yet?</a>
         <c:if test="${param.error!= null}">
             <span class="invalid">Invalid username and password</span>
         </c:if>
         <c:if test="${param.logout!= null}">
+            <br>
             <span class="invalid">You`ve been logged out</span>
         </c:if>
     </form>
-    <a href="/register">Not register yet?</a>
+
 </div>
+
+<script src="/js/login.js"></script>
 </body>
 </html>
